Use Dockerfile in docker compose; add `speedtest-demo` name
Rather than specifying an image, the `build` directive is now being used
in the docker compose project, so we are always running a container from
the sibling Dockerfile.

Also remove version number as it is obsolete, and add a `name` instead:
- https://docs.docker.com/compose/compose-file/04-version-and-name/

The docker compose config has been updated in the README too.