Clean apt-get cache after package installation
When installing packages an index is cached locally by default. The
index can be used on-the-fly and not stored locally. This change fixes
- `docker:S6587` - Cache should be cleaned after package installation

After making this change the image size increased slightly, but this is
due to the newer speedtest cli install. The layer installing curl etc.
has decreased from 49.9MB to 32.4MB.

More info on the problem:
 - https://rules.sonarsource.com/docker/RSPEC-6587/

Docker images should only contain necessary data. The package index is
optional for the correct working of the installed software. Storing an
index also increases the size of the Docker image. It should be reduced
to speed up deployments and reduce storage and bandwidth.

---

Other sonar rules for Dockerfile of interest
- https://community.sonarsource.com/t/12-new-rules-to-have-intentional-and-consistent-dockerfiles/102117