Add support for server flags (-s and -o)
We want to be able to pass a server ID or fully qualified hostname to
the speedtest CLI command so that we can use a specific server to test
against, rather than the default server that the CLI selects for us.

This change adds the new environment variable and a function to pass in
the hostname argument - `call_speedtest_with_args`.

This ultimately adds the `-o "$1"` argument to the `speedtest` CLI
command invocation:
- `JSON=$(speedtest --accept-license --accept-gdpr -f json -o "$1")`

If the new environment variable is not present, the CLI command will be
called as usual:
- `JSON=$(speedtest --accept-license --accept-gdpr -f json)`

---

To get the available ids, sponsors, and hostnames from speedtest run:
- `curl https://cli.speedtest.net/api/cli/config | jq -r '.servers[] | "id: \(.id), sponsor: \(.sponsor), host: \(.host)"'`

1. Build an image from the Dockerfile:
   - `docker build -t robinmanuelthiel/speedtest:latest .`
2. Simply remove the port number from the entry and run the container
   with the new environment variable:
   - `docker run -e SPEEDTEST_HOSTNAME=lg-tok.fdcservers.net robinmanuelthiel/speedtest:latest`

---

```shell
root@9d09012b3055:/# speedtest --help
Speedtest by Ookla is the official command line client for testing the speed and performance of your internet connection.

Version: speedtest 1.2.0.84

Usage: speedtest [<options>]
  -h, --help                        Print usage information
  -V, --version                     Print version number
  -L, --servers                     List nearest servers
  -s, --server-id=#                 Specify a server from the server list using its id
  -o, --host=ARG                    Specify a server, from the server list, using its host's fully qualified domain name
```

---

There is a new check added for the JSON response variable; if it is
invalid JSON the logging of the results and the db save action is
skipped:

```shell
if jq -e . >/dev/null 2>&1 <<<"$JSON"; then
   # do logging/db saving
  else
   echo "Unable to retrieve results from speedtest"
fi
```